better responsive
