+++ src/App.vue
@@ -4,7 +4,6 @@
   <div id="app">
 
     <NavBar />
-    <SideNav />
     <router-view />
   </div>
 
@@ -13,12 +12,11 @@
 <script>
 
   import NavBar from '@/components/NavBar'
-  import SideNav from '@/components/Sidenav'
+
   export default {
     name: 'App',
     components: {
-      NavBar,
-      SideNav
+      NavBar
     }
   }
